[Rename] 쿠폰 통합 테스트 통합 테스트 지원 파일 위치 이동으로 인한 수정
