+++ src/main/java/kr/hhplus/be/server/domain/coupon/repository/CouponRepository.java
@@ -4,6 +4,4 @@ import kr.hhplus.be.server.domain.coupon.entity.Coupon;
 
 public interface CouponRepository {
     Coupon findByIdWithLock(Long couponId);
-
-    void save(Coupon coupon);
 }
